+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - ALPHA: Added Calendar extension - shipped disabled
 - Added fallback for Mailfence and other non-quote-compliant providers - [#209](https://github.com/hkdb/aerion/issues/209)
 - Fixed CardDav remove provider code path to not leave orphaned contacts in db
+- Fixed Sent/Draft folder message listing - [#227](https://github.com/hkdb/aerion/issues/227)
 
 
 **v0.2.5 - 05-27-2026**